requirements.txt
python-telegram-bot==21.6
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.9.2
